+++ src/components/Charts/HashRates.tsx
