+++ src/components/Charts/HashRatesChart.tsx
