+++ src/components/Charts/POWChart.tsx
